+++ src/main/java/com/lubid/lubiduser/Repository/JwtMappingRepository.java
@@ -4,7 +4,7 @@ import com.lubid.lubiduser.model.JwtMappingUser;
 import com.lubid.lubiduser.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JwtMappingRepository extends JpaRepository<JwtMappingUser, Integer> {
+public interface JwtMappingRepository extends JpaRepository<JwtMappingUser, Long> {
 
     JwtMappingUser findByUser(User user);
