+++ src/main/java/com/lubid/lubiduser/Repository/UserAccountRepositiry.java
@@ -3,5 +3,5 @@ package com.lubid.lubiduser.Repository;
 import com.lubid.lubiduser.model.UserAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserAccountRepositiry extends JpaRepository<UserAccount, Integer> {
+public interface UserAccountRepositiry extends JpaRepository<UserAccount, Long> {
 }
